+++ src/components/sidebar/Sidebar.tsx
@@ -3,7 +3,7 @@ import { FaMapMarkerAlt } from 'react-icons/fa';
 import * as Styles from './styles';
 import { GpsButton } from '../GpsButton';
 import formatDate from '../../utils/functions/formatDate';
-import { useWeather } from '../../hooks/WeatherHook/weather';
+import { useWeather } from '../../hooks/WeatherHook';
 import { getIcon } from '../../utils/functions/weatherIcon';
 
 const Sidebar: React.FC = () => {
