+++ src/hooks/LoadingHook/index.tsx
@@ -1,7 +1,28 @@
-import React from 'react';
+import React, { createContext, useCallback, useState, useContext } from 'react';
 
-import { LoadingProvider } from './loading';
+interface LoadingContextProps {
+	isLoading: boolean;
+	setLoading: (loading: boolean) => void;
+}
 
-const LoadProvider: React.FC = ({ children }) => <LoadingProvider>{children}</LoadingProvider>;
+const LoadingContext = createContext<LoadingContextProps>({} as LoadingContextProps);
 
-export default LoadProvider;
+export const LoadingProvider: React.FC = ({ children }) => {
+	const [isLoading, setIsLoading] = useState<boolean>(false);
+
+	const setLoading = useCallback((loading: boolean) => {
+		setIsLoading(loading);
+	}, []);
+
+	return <LoadingContext.Provider value={{ isLoading, setLoading }}>{children}</LoadingContext.Provider>;
+};
+
+export function useLoading(): LoadingContextProps {
+	const context = useContext(LoadingContext);
+
+	if (!context) {
+		throw new Error('useLoading must be used within an LoadingProvider');
+	}
+
+	return context;
+}
