+++ src/hooks/index.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import { WeatherProvider } from './WeatherHook';
+import { LoadingProvider } from './LoadingHook';
+
+const AppProvider: React.FC = ({ children }) => {
+	return (
+		<LoadingProvider>
+			<WeatherProvider>{children}</WeatherProvider>
+		</LoadingProvider>
+	);
+};
+
+export default AppProvider;
