+++ src/pages/Home.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { InitalHomePage } from '../components/Loading';
-import { useWeather } from '../hooks/WeatherHook/weather';
-import { useLoading } from '../hooks/LoadingHook/loading';
+import { useWeather } from '../hooks/WeatherHook';
+import { useLoading } from '../hooks/LoadingHook';
 import { requestByLattAndLong } from '../services/api.requests';
 
 const Home: React.FC = ({ children }) => {
@@ -10,14 +10,12 @@ const Home: React.FC = ({ children }) => {
 
 	React.useEffect(() => {
 		const exec = async (): Promise<void> => {
-			console.log('here');
 			setLoading(true);
 			const nearstCities = await requestByLattAndLong();
 			if (nearstCities.length > 0) {
 				getClimate(nearstCities[0].woeid);
 			}
 			setLoading(false);
-			console.log('here');
 		};
 
 		exec();
